ml-lens/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
anthropic>=0.39.0
openai>=1.30.0
pymupdf>=1.24.0
arxiv>=2.1.0
httpx>=0.27.2
e2b>=2.20.0
langfuse>=2.53.1
